contains setOnClickListener, setOnKeyListener & TextWather templates realized adding portions to daily ration
implementation of material design

+++ app/src/main/java/ru/xpendence/development/gimstopwatch/foodstuffs/GoodInDayRation.java
@@ -53,7 +53,7 @@ public class GoodInDayRation {
     }
 
     public void setCalories(int calories, int amount) {
-        this.calories = calories / 100 * amount;
+        this.calories = (int) ((calories / 100d) * amount);
     }
 
     public String getCategory() {
